wip:code clean up
