+++ server/build.gradle.kts
@@ -19,7 +19,6 @@ dependencies {
     implementation(projects.shared)
 
     implementation(libs.logback)
-    implementation(libs.logback)
 
     implementation(libs.ktor.server.core.jvm)
     implementation(libs.ktor.server.netty)
