+++ shared/src/commonMain/composeResources/values/version.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_version">39.0.5 (67)</string>
+    <string name="app_version">1.0.0</string>
 </resources>
