--- shared/src/commonMain/kotlin/org/example/project/SharedSDK.kt
@@ -1,21 +1,0 @@
-package org.example.project
-
-import org.example.project.data.response.AuthResponse
-import org.example.project.netio.APIService
-
-object SharedSDK {
-
-    //    private val localDb = Database(getDatabaseDriverFactory())
-    val remoteApi = APIService()
-
-    //    suspend fun saveToken(data: AuthResponse) = localDb.addUserSetting(data)
-    suspend fun saveToken(data: AuthResponse) = "maderchod k bacche"
-    suspend fun getToken(): String? {
-//        val userSetting = localDb.getUserSetting().filter {
-//            it.settingkey == "token"
-//        }
-//        return userSetting.firstOrNull()?.settingvalue
-        return "lund SQL"
-    }
-
-}
