--- shared/src/commonMain/kotlin/org/example/project/data/KMPConstant.kt
@@ -1,10 +1,0 @@
-package org.example.project.data
-
-object KMPConstant {
-
-    const val SUCCESS_MESSAGE = "Success"
-    const val ERROR_MESSAGE_1 = "Check Email and password length"
-    const val ERROR_MESSAGE_2 = "Unable to create Account"
-    const val ERROR_MESSAGE_3 = "Incorrect username or password"
-    const val ERROR_MESSAGE_4 = "Bad request dear"
-}
