+++ shared/src/commonMain/kotlin/org/example/project/shared.kt
@@ -1,6 +1,3 @@
 package org.example.project
 
-val SERVER_PORT = 8080
-val GREETING = "Hello World!"
-
-//expect fun getPlatformName(): String
+expect fun getPlatformName(): String
