+++ shared/src/commonMain/kotlin/org/example/project/ui/screens/home/chat/ChatListScreen.kt
@@ -7,40 +7,12 @@ import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import org.koin.compose.viewmodel.koinViewModel
 import org.example.project.ui.components.FullScreenError
 import org.example.project.ui.components.ItemChat
 import org.example.project.ui.components.SimpleLoading
-
-//
-//class ChatListScreen : Screen {
-//
-//    @Composable
-//    override fun Content() {
-//
-//        val chatViewModel =
-//            getViewModel(ChatListScreen().key, viewModelFactory { ChatViewModel() })
-//
-//        val navigator = LocalNavigator.currentOrThrow
-//
-//        LaunchedEffect(Unit) {
-//            println("UI-Compose : ChatListScreen")
-//            chatViewModel.getOnlineUser()
-//        }
-//
-//        ChatListScreenContent(
-//            uiState = chatViewModel.onlineUserChatListUiState,
-//            onUserSelect = {
-//                navigator.push(
-//                    ChatDetailScreen(it)
-//                )
-//            }
-//        )
-//
-//    }
+import org.koin.compose.viewmodel.koinViewModel
 
 @Composable
 fun ChatListScreen(
