+++ shared/src/commonMain/kotlin/org/example/project/ui/screens/home/profile/ProfileViewModel.kt
@@ -1,20 +1,20 @@
 package org.example.project.ui.screens.home.profile
 
-import org.example.project.SharedSDK
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.launch
+import org.example.project.AppService
 
-class ProfileViewModel : ViewModel() {
+class ProfileViewModel(
+    private val service: AppService
+) : ViewModel() {
 
     companion object {
         const val TAG = "ProfileViewModel"
     }
-
-    private val sdk = SharedSDK
 
     var onLineUiState by mutableStateOf(false)
         private set
@@ -22,7 +22,7 @@ class ProfileViewModel : ViewModel() {
     init {
         println("$TAG init")
         viewModelScope.launch {
-            onLineUiState = sdk.remoteApi.isUserConnected()
+            onLineUiState = service.isUserConnected()
         }
     }
