+++ shared/src/iosMain/kotlin/org/example/project/common.ios.kt
@@ -1,15 +1,15 @@
-package org.example.project
-
-import com.squareup.sqldelight.db.SqlDriver
-import com.squareup.sqldelight.drivers.native.NativeSqliteDriver
-import org.example.project.db.AppDatabase
-
-actual class DatabaseDriverFactory {
-    actual fun createDriver(): SqlDriver {
-        return NativeSqliteDriver(AppDatabase.Schema, "test.db")
-    }
-}
-
-actual fun getDatabaseDriverFactory() = DatabaseDriverFactory()
-
-actual fun getPlatformName(): String = "iOS"
+//package org.example.project
+//
+//import com.squareup.sqldelight.db.SqlDriver
+//import com.squareup.sqldelight.drivers.native.NativeSqliteDriver
+//import org.example.project.db.AppDatabase
+//
+//actual class DatabaseDriverFactory {
+//    actual fun createDriver(): SqlDriver {
+//        return NativeSqliteDriver(AppDatabase.Schema, "test.db")
+//    }
+//}
+//
+//actual fun getDatabaseDriverFactory() = DatabaseDriverFactory()
+//
+//actual fun getPlatformName(): String = "iOS"
