+++ shared/src/jvmMain/kotlin/org/example/project/shared.jvm.kt
@@ -1,15 +1,16 @@
 package org.example.project
 
 import com.squareup.sqldelight.db.SqlDriver
-import org.example.project.db.AppDatabase
 
-actual fun getPlatformName(): String = "Server"
+//import org.example.project.db.AppDatabase
+
+actual fun getPlatformName(): String = "JVM"
 
 
-actual class DatabaseDriverFactory {
-    actual fun createDriver(): SqlDriver {
-        return NativeSqliteDriver(AppDatabase.Schema, "test.db")
-    }
-}
-
-actual fun getDatabaseDriverFactory() = DatabaseDriverFactory()
+//actual class DatabaseDriverFactory {
+//    actual fun createDriver(): SqlDriver {
+//        return NativeSqliteDriver(AppDatabase.Schema, "test.db")
+//    }
+//}
+//
+//actual fun getDatabaseDriverFactory() = DatabaseDriverFactory()
